Change MainCut to decimal, update DesignList handling
Changed MainCut property from int? to decimal? in both ProgramEntry and ProgramVM to support fractional values. Updated ProgramController to pass raw _db.Designs to the view, and moved SelectList creation for DesignList to AddEdit.cshtml. No functional changes to GetMatchingByDesign.

+++ src/Inventory.Domain/ViewModels/ProgramVM.cs
@@ -25,7 +25,7 @@ namespace Inventory.Domain.ViewModels
         public string? Remarks { get; set; }
         public string? Rate { get; set; }
         public DateOnly Date { get; set; }
-        public int? MainCut { get; set; }
+        public decimal? MainCut { get; set; }
         public int? Quantity { get; set; }
         public int? Round { get; set; }
